+++ src/lib/tasks/task.ts
@@ -6,14 +6,14 @@ export interface SimpleGitTaskConfiguration<RESPONSE, FORMAT, INPUT> {
 
    commands: string[]
    format: FORMAT;
-   parser (text: INPUT): RESPONSE;
+   parser(text: INPUT): RESPONSE;
 
    concatStdErr?: boolean;
    onError?: (
       exitCode: number,
       error: string,
-      done:(result: INPUT) => void,
-      fail:(error: string) => void,
+      done: (result: INPUT) => void,
+      fail: (error: string) => void,
    ) => void;
 
 }
@@ -38,17 +38,17 @@ export function configurationErrorTask(error: Error | string): EmptyTask {
    }
 }
 
-export function straightThroughStringTask (commands: string[]): StringTask<string> {
+export function straightThroughStringTask(commands: string[]): StringTask<string> {
    return {
       commands,
       format: 'utf-8',
-      parser (text) {
+      parser(text) {
          return text;
       },
    }
 }
 
-export function isBufferTask<R> (task: SimpleGitTask<R>): task is BufferTask<R> {
+export function isBufferTask<R>(task: SimpleGitTask<R>): task is BufferTask<R> {
    return task.format === 'buffer';
 }